change allocatedtime to doubles + other fixes

See merge request xnadzam/pv-168-task-manager!100

+++ src/main/java/cz/muni/fi/pv168/project/business/model/LogTimeInfo.java
@@ -6,28 +6,28 @@ import cz.muni.fi.pv168.project.business.model.abstracts.Entity;
  * @author Vladimir Borek
  */
 public class LogTimeInfo extends Entity {
-    private Integer loggedTime;
+    private Double loggedTime;
     private final User user;
     private final Long taskID;
 
-    public LogTimeInfo(Integer loggedTime, User user, Long taskID){
+    public LogTimeInfo(Double loggedTime, User user, Long taskID){
         super(null);
         this.loggedTime = loggedTime;
         this.user = user;
         this.taskID = taskID;
     }
 
-    public LogTimeInfo(Long id, Integer loggedTime, User user, Long taskID){
+    public LogTimeInfo(Long id, Double loggedTime, User user, Long taskID){
         super(id);
         this.loggedTime = loggedTime;
         this.user = user;
         this.taskID = taskID;
     }
 
-    public Integer getLoggedTime(){
+    public Double getLoggedTime(){
         return this.loggedTime;
     }
-    public void setLoggedTime(Integer newLoggedTime){
+    public void setLoggedTime(Double newLoggedTime){
         this.loggedTime = newLoggedTime;
     }
